@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, DoCheck } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatIconModule } from '@angular/material/icon';
@@ -7,9 +7,10 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { RouterOutlet } from '@angular/router';
 import { finalize } from 'rxjs';
 
-import { Golden } from './golden';
+import { MotionGolden } from './golden';
 import { GoldensService } from './goldens.service';
 import { TestOverviewComponent } from './test-overview/test-overview.component';
+import { ProgressTracker } from '../utils/progress';
 
 @Component({
   selector: 'app-root',
@@ -21,39 +22,43 @@ import { TestOverviewComponent } from './test-overview/test-overview.component';
     MatIconModule,
     MatDividerModule,
     MatButtonModule,
+    MatDividerModule,
     MatProgressBarModule,
   ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })
-export class AppComponent {
-  constructor(private goldenService: GoldensService) {}
+export class AppComponent implements DoCheck {
+  constructor(
+    private goldenService: GoldensService,
+    private progressTracker: ProgressTracker,
+  ) {}
 
-  _progressTracker = 0;
-
-  goldens: Golden[] = [];
-
-  get showProgress() {
-    return this._progressTracker > 0;
+  ngDoCheck() {
+    this.showProgress = this.progressTracker.isActive;
   }
+
+  goldens: MotionGolden[] = [];
+
+  showProgress = false;
 
   ngOnInit(): void {
     this.fetchGoldens();
   }
 
   fetchGoldens(): void {
-    this._progressTracker++;
+    this.progressTracker.beginProgress();
     this.goldenService
       .getGoldens()
-      .pipe(finalize(() => this._progressTracker--))
+      .pipe(finalize(() => this.progressTracker.endProgress()))
       .subscribe((goldens) => (this.goldens = goldens));
   }
 
   refreshGoldens(clear: boolean): void {
-    this._progressTracker++;
+    this.progressTracker.beginProgress();
     this.goldenService
       .refreshGoldens(clear)
-      .pipe(finalize(() => this._progressTracker--))
+      .pipe(finalize(() => this.progressTracker.endProgress()))
       .subscribe((goldens) => (this.goldens = goldens));
   }
 }
